KmgFundGenMessageTypes、KmgFundLogMessageTypes、KmgFundValMessageTypesを新規作成し、KmgYamlUtilsのインポート文を修正しました。
