+++ src/main/java/kmg/fund/infrastructure/types/msg/KmgFundGenMessageTypes.java
@@ -1,15 +1,15 @@
-package kmg.fund.domain.types.msg;
+package kmg.fund.infrastructure.types.msg;
 
 import java.util.HashMap;
 import java.util.Map;
 
 import kmg.fund.infrastructure.common.KmgFundComExcMessageTypes;
-import kmg.fund.infrastructure.common.KmgFundComValMessageTypes;
+import kmg.fund.infrastructure.common.KmgFundComGenMessageTypes;
 
 /**
- * KMG 基盤バリデーションメッセージの種類<br>
+ * KMG 基盤一般メッセージの種類<br>
  * <p>
- * Fundは、Foundationの略。
+ * Msgは、Messageの略。
  * </p>
  *
  * @author KenichiroArai
@@ -18,7 +18,7 @@ import kmg.fund.infrastructure.common.KmgFundComValMessageTypes;
  *
  * @version 0.1.0
  */
-public enum KmgFundValMessageTypes implements KmgFundComValMessageTypes, KmgFundComExcMessageTypes {
+public enum KmgFundGenMessageTypes implements KmgFundComGenMessageTypes, KmgFundComExcMessageTypes {
 
     /* 定義：開始 */
 
@@ -33,6 +33,28 @@ public enum KmgFundValMessageTypes implements KmgFundComValMessageTypes, KmgFund
      */
     NONE("指定無し"),
 
+    /**
+     * 該当するYAMLファイルがありません。ファイルパス=[{0}]
+     *
+     * @author KenichiroArai
+     *
+     * @since 0.1.0
+     *
+     * @version 0.1.0
+     */
+    KMGFUND_GEN24000("該当するYAMLファイルがありません。ファイルパス=[{0}]"),
+
+    /**
+     * YAMLにロードするファイルの読み込みに失敗しました。ファイルパス=[{0}]
+     *
+     * @author KenichiroArai
+     *
+     * @since 0.1.0
+     *
+     * @version 0.1.0
+     */
+    KMGFUND_GEN24001("YAMLにロードするファイルの読み込みに失敗しました。ファイルパス=[{0}]"),
+
     /* 定義：終了 */
 
     ;
@@ -42,14 +64,14 @@ public enum KmgFundValMessageTypes implements KmgFundComValMessageTypes, KmgFund
      *
      * @since 0.1.0
      */
-    private static final Map<String, KmgFundValMessageTypes> VALUES_MAP = new HashMap<>();
+    private static final Map<String, KmgFundGenMessageTypes> VALUES_MAP = new HashMap<>();
 
     static {
 
         /* 種類のマップにプット */
-        for (final KmgFundValMessageTypes type : KmgFundValMessageTypes.values()) {
+        for (final KmgFundGenMessageTypes type : KmgFundGenMessageTypes.values()) {
 
-            KmgFundValMessageTypes.VALUES_MAP.put(type.get(), type);
+            KmgFundGenMessageTypes.VALUES_MAP.put(type.get(), type);
 
         }
 
@@ -90,9 +112,9 @@ public enum KmgFundValMessageTypes implements KmgFundComValMessageTypes, KmgFund
      *
      * @return デフォルト値
      */
-    public static KmgFundValMessageTypes getDefault() {
+    public static KmgFundGenMessageTypes getDefault() {
 
-        final KmgFundValMessageTypes result = NONE;
+        final KmgFundGenMessageTypes result = NONE;
         return result;
 
     }
@@ -110,9 +132,9 @@ public enum KmgFundValMessageTypes implements KmgFundComValMessageTypes, KmgFund
      *
      * @return 種類。指定無し（NONE）：キーが存在しない場合。
      */
-    public static KmgFundValMessageTypes getEnum(final String key) {
+    public static KmgFundGenMessageTypes getEnum(final String key) {
 
-        KmgFundValMessageTypes result = KmgFundValMessageTypes.VALUES_MAP.get(key);
+        KmgFundGenMessageTypes result = KmgFundGenMessageTypes.VALUES_MAP.get(key);
 
         if (result == null) {
 
@@ -130,9 +152,9 @@ public enum KmgFundValMessageTypes implements KmgFundComValMessageTypes, KmgFund
      *
      * @return 初期値
      */
-    public static KmgFundValMessageTypes getInitValue() {
+    public static KmgFundGenMessageTypes getInitValue() {
 
-        final KmgFundValMessageTypes result = NONE;
+        final KmgFundGenMessageTypes result = NONE;
         return result;
 
     }
@@ -145,7 +167,7 @@ public enum KmgFundValMessageTypes implements KmgFundComValMessageTypes, KmgFund
      * @param displayName
      *                    表示名
      */
-    KmgFundValMessageTypes(final String displayName) {
+    KmgFundGenMessageTypes(final String displayName) {
 
         this.displayName = displayName;
         this.key = super.name();
